refactor: 로또 숫자 생성 fixture 클래스 패키지 이동
